Rename to FlamethrowerProjectile
