+++ ExampleMod/Projectiles/FlamethrowerProjectile.png
